requirements.txt
flask>=2.0.0
google-genai>=1.0.0
Pillow>=9.0.0
python-dotenv>=1.0.0
